+++ Sofa/Component/Constraint/Lagrangian/Solver/src/sofa/component/constraint/lagrangian/solver/ConstraintSolverImpl.h
@@ -93,6 +93,10 @@ public:
 
     void removeConstraintCorrection(core::behavior::BaseConstraintCorrection *s) override;
 
+    MultiLink< ConstraintSolverImpl,
+        core::behavior::BaseConstraintCorrection,
+        BaseLink::FLAG_STOREPATH> l_constraintCorrections;
+
 protected:
 
     void postBuildSystem(const core::ConstraintParams* cParams) override;
@@ -100,9 +104,6 @@ protected:
 
     void clearConstraintCorrections();
 
-    MultiLink< ConstraintSolverImpl,
-        core::behavior::BaseConstraintCorrection,
-        BaseLink::FLAG_STOREPATH> l_constraintCorrections;
 
     /// Calls the method resetConstraint on all the mechanical states and BaseConstraintSet
     /// In the case of a MechanicalObject, it clears the constraint jacobian matrix
